# openapi/users.yaml
openapi: 3.0.3
info:
  title: Users Service API
  version: 1.0.0
  description: API del microservicio de usuarios.
servers:
  - url: https://g7-riesgo-ansiedad-users-gzamfsdbdxh8adh4.chilecentral-01.azurewebsites.net
    description: Servicio de usuarios desplegado en Azure
tags:
  - name: Users
    description: Operaciones de usuarios
  - name: Health
    description: Estado del servicio y su base de datos
paths:
  /users:
    post:
      tags:
        - Users
      summary: Crear un usuario
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            example:
              names: Ana
              surnames: Perez
              code: EST-12345
              email: ana.perez@example.com
              password: "UnaClaveSegura123!"
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Los datos enviados no son válidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: El email o el código de estudiante ya está registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageError"
              example:
                message: El email ya está registrado
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageError"
  /health:
    get:
      tags:
        - Health
      summary: Consultar el estado del servicio
      operationId: getHealth
      responses:
        "200":
          description: El servicio y la base de datos están disponibles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
              example:
                service: users
                status: OK
                database: OK
        "503":
          description: La base de datos no está disponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
              example:
                service: users
                status: ERROR
                database: UNAVAILABLE
components:
  schemas:
    CreateUserRequest:
      type: object
      required:
        - names
        - surnames
        - code
        - email
        - password
      properties:
        names:
          type: string
          minLength: 1
          description: Nombres del usuario
        surnames:
          type: string
          minLength: 1
          description: Apellidos del usuario
        code:
          type: string
          minLength: 1
          description: Código de estudiante único
        email:
          type: string
          format: email
          description: Correo electrónico único
        password:
          type: string
          minLength: 1
          format: password
          writeOnly: true
          description: Contraseña del usuario. No se devuelve en la respuesta.
    User:
      type: object
      required:
        - id
        - names
        - surnames
        - code
        - email
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        names:
          type: string
        surnames:
          type: string
        code:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    MessageError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
              message:
                type: string
    HealthStatus:
      type: object
      required:
        - service
        - status
        - database
      properties:
        service:
          type: string
          enum:
            - users
        status:
          type: string
          enum:
            - OK
            - ERROR
        database:
          type: string
          enum:
            - OK
            - UNAVAILABLE